add in email link intention and check for redirect
